test(snapshot): wait out the ETXTBSY window for every stand-in binary
`a_live_dm_device_is_not_removed` failed on Linux CI with "Text file
busy": a sibling test thread forked between this thread's create and
close of the fake dmsetup, so its child held a write fd and the kernel
refused the exec. The block-tools stand-ins already retried the exec for
exactly this reason; `fake_dmsetup` did not, and this PR's new tests fork
often enough to make that window matter.

The retry now lives in one place — a crate-level `test_support` both
paths use — so the next stand-in inherits it instead of rediscovering the
flake. The probe run's own log line is cleared, since the dmsetup
stand-in records every invocation the test then reads back.
