+++ engine/arcbox-snapshot/src/snapshot_cow/block_tools/mod.rs
@@ -227,78 +227,3 @@ mod tests {
         }
     }
 }
-
-/// Shared scaffolding for the two implementations' tests: a stand-in for
-/// the real binary that records how it was invoked.
-#[cfg(test)]
-mod test_support {
-    use std::path::{Path, PathBuf};
-    use std::process::Command;
-    use std::time::Duration;
-
-    /// Write an executable stand-in for `name` into `dir`: a shell script
-    /// that logs every invocation's arguments to `<name>.calls` and runs
-    /// `body` (a `case "$*"` over them). Returns the script and the call
-    /// log.
-    ///
-    /// Tests hand these scripts `/dev/loop9999`, a device no host has, so
-    /// the sysfs verification finds nothing to compare and the script stays
-    /// the only authority on what happened.
-    pub(super) fn fake_tool(dir: &Path, name: &str, body: &str) -> (PathBuf, PathBuf) {
-        use std::os::unix::fs::PermissionsExt as _;
-
-        let calls = dir.join(format!("{name}.calls"));
-        let script = dir.join(name);
-        std::fs::write(
-            &script,
-            format!(
-                "#!/bin/sh\nprintf '%s\\n' \"$*\" >> {}\ncase \"$*\" in\n{body}\nesac\n",
-                calls.display()
-            ),
-        )
-        .unwrap();
-        std::fs::set_permissions(&script, std::fs::Permissions::from_mode(0o755)).unwrap();
-        wait_until_executable(&script);
-        // The probe ran the script, so every test starts from an empty record.
-        std::fs::remove_file(&calls).ok();
-        (script, calls)
-    }
-
-    /// Every line the stand-in recorded, in invocation order.
-    pub(super) fn calls(log: &Path) -> Vec<String> {
-        std::fs::read_to_string(log)
-            .unwrap_or_default()
-            .lines()
-            .map(str::to_owned)
-            .collect()
-    }
-
-    /// Run the freshly written `script` until the kernel stops refusing it.
-    ///
-    /// A sibling test thread that forks — `CowManager::new` probes every
-    /// `dmsetup` candidate, so most of this crate's tests do — between this
-    /// thread's `create` and `close` leaves its child holding a write fd to
-    /// the script, and Linux will not exec a file that is open for writing
-    /// (`ETXTBSY`). The window closes as soon as that child execs and never
-    /// reopens, since nothing writes the script again; exec'ing it until it
-    /// runs is what proves the window is shut.
-    fn wait_until_executable(script: &Path) {
-        let deadline = std::time::Instant::now() + Duration::from_secs(5);
-        loop {
-            match Command::new(script).output() {
-                Err(error) if error.kind() == std::io::ErrorKind::ExecutableFileBusy => {
-                    assert!(
-                        std::time::Instant::now() < deadline,
-                        "{} stayed busy for 5s",
-                        script.display()
-                    );
-                    std::thread::sleep(Duration::from_millis(5));
-                }
-                other => {
-                    other.unwrap();
-                    return;
-                }
-            }
-        }
-    }
-}
